refactor: 💡 addressed pr comments
