+++ src/declarations/marketplace-v2.d.ts
@@ -78,12 +78,14 @@ export type ListingStatus =
 export type NFTStandard = { DIP721v2: null };
 export interface Offer {
   status: OfferStatus;
-  token_id: string;
+  token_id: bigint;
   time: bigint;
   fungible_id: Principal;
   token_owner: Principal;
   buyer: Principal;
   price: bigint;
+  tokenId: bigint;
+  created: bigint;
 }
 export type OfferStatus =
   | { Bought: null }
